Crudely move sagas to the main process
