--- src/hooks/useServerValidation.js
@@ -1,4 +1,0 @@
-import { useCallableSaga } from '../components/SagaMiddlewareProvider';
-import { validateServerUrl } from '../sagas/servers';
-
-export const useServerValidation = () => useCallableSaga(validateServerUrl, []);
