+++ src/main/sagas/index.js
@@ -1,4 +1,5 @@
-import { spawn, call, take } from 'redux-saga/effects';
+import { spawn, call, take, takeEvery, select } from 'redux-saga/effects';
+import { createStructuredSelector } from 'reselect';
 
 import { dockSaga } from './dock';
 import { trayIconSaga } from './trayIcon';
@@ -12,6 +13,9 @@ import { deepLinksSaga } from './deepLinks';
 import { navigationSaga } from './navigation';
 import { updatesSaga } from './updates';
 import { spellCheckingSaga } from './spellChecking';
+import { preferencesSaga } from './preferences';
+import { serversSaga } from './servers';
+import { writeToStorage } from '../localStorage';
 
 export function *rootSaga() {
 	yield take(appReadyChannel());
@@ -20,12 +24,34 @@ export function *rootSaga() {
 	const rootWindow = yield call(rootWindowSaga);
 
 	yield spawn(appSaga, rootWindow);
+	yield spawn(preferencesSaga, rootWindow);
+	yield spawn(serversSaga, rootWindow);
 	yield spawn(deepLinksSaga);
-	yield spawn(navigationSaga);
-	yield spawn(updatesSaga);
+	yield spawn(navigationSaga, rootWindow);
+	yield spawn(updatesSaga, rootWindow);
 	yield spawn(spellCheckingSaga, rootWindow);
 	yield spawn(menuBarSaga, rootWindow);
 	yield spawn(touchBarSaga, rootWindow);
 	yield spawn(dockSaga);
 	yield spawn(trayIconSaga);
+
+	const selectPersistableValues = createStructuredSelector({
+		currentServerUrl: ({ currentServerUrl }) => currentServerUrl ?? null,
+		doCheckForUpdatesOnStartup: ({ doCheckForUpdatesOnStartup }) => doCheckForUpdatesOnStartup ?? true,
+		isMenuBarEnabled: ({ isMenuBarEnabled }) => isMenuBarEnabled ?? true,
+		isShowWindowOnUnreadChangedEnabled: ({ isShowWindowOnUnreadChangedEnabled }) => isShowWindowOnUnreadChangedEnabled ?? false,
+		isSideBarEnabled: ({ isSideBarEnabled }) => isSideBarEnabled ?? true,
+		isTrayIconEnabled: ({ isTrayIconEnabled }) => isTrayIconEnabled ?? true,
+		mainWindowState: ({ mainWindowState }) => mainWindowState ?? {},
+		servers: ({ servers }) => servers ?? [],
+		skippedUpdateVersion: ({ skippedUpdateVersion }) => skippedUpdateVersion ?? null,
+		trustedCertificates: ({ trustedCertificates }) => trustedCertificates ?? {},
+	});
+
+	yield takeEvery('*', function *() {
+		const values = yield select(selectPersistableValues);
+		for (const [key, value] of Object.entries(values)) {
+			yield call(writeToStorage, rootWindow, key, value);
+		}
+	});
 }
